+++ Sources/Utils.swift
@@ -30,9 +30,11 @@ public enum LogLevel: Int {
 
 public var logLevel: LogLevel = .Errors
 
+public var logger: (LogLevel, Any) -> Void = { print($1) }
+
 func log(level logLevel: LogLevel, _ message: Any) {
   guard logLevel.rawValue <= Dip.logLevel.rawValue else { return }
-  print(message)
+  logger(logLevel, message)
 }
 
 ///Internal protocol used to unwrap optional values.
@@ -70,10 +72,10 @@ class WeakBox<T>: WeakBoxType {
   }
 
   init(_ value: T) {
-    #if !_runtime(_ObjC) || !swift(>=3.0)
-      weak var value: AnyObject? = value as? AnyObject
-    #else
+    #if _runtime(_ObjC)
       weak var value: AnyObject? = value as AnyObject
+    #else
+      weak var value: AnyObject? = value as? AnyObject
     #endif
     guard value != nil else {
       fatalError("Can not store weak reference to not a class instance (\(T.self))")